requirements.txt
python-dotenv
fastapi
uvicorn
langchain
langgraph
langchain-google-genai
uvicorn
pydantic
python-dotenv
langchain-core
duckduckgo-search
httpx
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pandas